Image files will now be deleted from assets folder when replaced with other images while editing. Also fixed issue where placeholder image was being deleted because of this process. About to work on implementing this in the delete section, but will probably upload a sprint before I do so, because it's getting quite late.

+++ StageBear/Controllers/ShowsController.cs
@@ -57,7 +57,7 @@ namespace StageBear.Controllers
                 }
                 else
                 {
-                    show.Image = "ShakesLogo.png";
+                    show.Image = "ShakesPlaceholder.png";
                 }
 
                     _context.Add(show);
@@ -128,7 +128,7 @@ namespace StageBear.Controllers
                         if(!string.IsNullOrEmpty(existingShow.Image))
                         {
                             string oldImagePath = Path.Combine(Directory.GetCurrentDirectory(),"wwwroot","assets",existingShow.Image);
-                            if (System.IO.File.Exists(oldImagePath))
+                            if (System.IO.File.Exists(oldImagePath) && existingShow.Image != "ShakesPlaceholder.png")
                             {
                                 System.IO.File.Delete(oldImagePath);
                             }
